% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAdbImport-import.R
\name{import.tRNAdb}
\alias{import.tRNAdb}
\alias{import.tRNAdb.id}
\alias{import.mttRNAdb.id}
\alias{import.tRNAdb.blast}
\alias{import.mttRNAdb.blast}
\alias{import.tRNAdb}
\alias{import.mttRNAdb}
\title{Importing information from the tRNA db as GRanges object}
\usage{
import.tRNAdb.id(tdbID, database = "DNA", origin = "allothers",
  dbURL = TRNA_DB_URL, verbose = FALSE)

import.mttRNAdb.id(tdbID, database = "DNA", origin = "allothers",
  dbURL = TRNA_DB_URL_MT, verbose = FALSE)

import.tRNAdb.blast(blastSeq, database = "DNA", origin = "allothers",
  dbURL = TRNA_DB_URL, verbose = FALSE)

import.mttRNAdb.blast(blastSeq, database = "DNA", origin = "allothers",
  dbURL = TRNA_DB_URL_MT, verbose = FALSE)

import.tRNAdb(organism = "", strain = "", taxonomyID = "",
  aminoacids = "", anticodons = "", sequences = list(),
  structures = list(), reference = "", comment = "", pubmed = "",
  genes = "", database = "DNA", origin = "allothers",
  dbURL = TRNA_DB_URL, verbose = FALSE)

import.mttRNAdb(organism = "", strain = "", taxonomyID = "",
  aminoacids = "", anticodons = "", sequences = list(),
  structures = list(), reference = "", comment = "", pubmed = "",
  genes = "", database = "DNA", origin = "allothers",
  dbURL = TRNA_DB_URL_MT, verbose = FALSE)
}
\arguments{
\item{tdbID}{a tRNAdb ID}

\item{database}{"RNA" or "DNA"}

\item{origin}{"plastid", "mitochondrial" or "allothers"}

\item{dbURL}{the URL of the tRNA db}

\item{verbose}{whether to report verbose information from the httr calls}

\item{blastSeq}{a sequence to use for a blast search}

\item{organism}{a organism name as a character string}

\item{strain}{a strain information as a character string}

\item{taxonomyID}{organism and strain information as a taxonom ID}

\item{aminoacids}{a character vector of amino acids as a three letter code}

\item{anticodons}{a character vector of anticodon sequences}

\item{sequences}{a named (1-15) list of sequences, which are used for the
search}

\item{structures}{a named (1-15) list of structures, which are used for the
search. Please use the \code{\(\)} or \code{><} dot bracket annotation.}

\item{reference}{a reference as a character string}

\item{comment}{a comment as a character string}

\item{pubmed}{a pubmed ID}

\item{genes}{a gene name as a character string}
}
\value{
a GRanges object containing the information from the tRNA db
}
\description{
title
}
\examples{
\donttest{
import.tRNAdb(organism = "Saccharomyces cerevisiae",
              aminoacids = c("Phe","Ala"),
              anticodons = c("GAA"))
import.tRNAdb.id(tdbID = "tdbD00000785")
import.tRNAdb.blast(blastSeq = "GCGGATTTAGCTCAGTTGGGAGAGCGCCAGACTGAAGATCTGGAGGTCCTGTGTTCGATCCACAGAATTCGCA")
}
}
