% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAdbImport-checks.R
\docType{methods}
\name{istRNAdbGRanges}
\alias{istRNAdbGRanges}
\alias{istRNAdbGRanges}
\alias{istRNAdbGRanges,GRanges-method}
\title{tRNAscan compatibility check}
\usage{
istRNAdbGRanges(gr)

\S4method{istRNAdbGRanges}{GRanges}(gr)
}
\arguments{
\item{gr}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{istRNAdbGRanges} checks whether a GRanges object contains the 
information expected for a tRNAscan result.
}
\examples{
\donttest{
istRNAdbGRanges(gr)
}
}
