% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tRNAdbImport-checks.R
\name{istRNAdbGRanges}
\alias{istRNAdbGRanges}
\alias{istRNAdbGRanges,GRanges-method}
\title{tRNAdb compatibility check}
\usage{
istRNAdbGRanges(x)

\S4method{istRNAdbGRanges}{GRanges}(x)
}
\arguments{
\item{x}{the \code{GRanges} object to test}
}
\value{
a logical value
}
\description{
\code{istRNAdbGRanges} checks whether a GRanges object contains the 
information expected for a tRNAdb result.
}
\examples{
gr <- import.tRNAdb(organism = "Saccharomyces cerevisiae",
              aminoacids = c("Phe","Ala"),
              anticodons = c("GAA"))
istRNAdbGRanges(gr)
}
