% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAdbImport.R
\docType{package}
\name{tRNAdbImport}
\alias{tRNAdbImport}
\alias{tRNAdbImport-package}
\title{tRNAdbImport: Importing from to tRNAdb and mitotRNAdb as GRanges}
\description{
The tRNAdb and mttRNAdb (Jühling et al. 2009) is a compilation of tRNA 
sequences and tRNA genes. It is a follow up version of the database of 
Sprinzl et al. 2005.

Using `tRNAdbImport` the tRNAdb can be accessed as outlined on the website 
[http://trna.bioinf.uni-leipzig.de/](http://trna.bioinf.uni-leipzig.de/) and 
the results are returned as a `GRanges` object.
}
\section{Manual}{

Please refer to the tRNAdbImport vignette for an example how to work and 
use the package: \href{../doc/tRNAdbImport.html}{tRNAdbImport}
}

\references{
Jühling F, Mörl M, Hartmann RK, Sprinzl M, Stadler PF, Pütz J. 2009. "tRNAdb 
2009: compilation of tRNA sequences and tRNA genes." Nucleic Acids Research, 
Volume 37 (suppl_1): D159–162. doi:10.1093/nar/gkn772.
}
\seealso{
[import.tRNAdb()] for examples
}
\author{
Felix G M Ernst [aut]
}
