% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAdbImport.R
\docType{package}
\name{tRNAdbImport}
\alias{tRNAdbImport}
\alias{tRNAdbImport-package}
\alias{TRNA_DB_URL}
\alias{TRNA_DB_URL_MT}
\title{tRNAdbImport: Importing from to tRNAdb and mitotRNAdb as GRanges}
\format{An object of class \code{character} of length 1.}
\usage{
TRNA_DB_URL

TRNA_DB_URL_MT
}
\description{
title
}
\references{
Jühling F, Mörl M, Hartmann RK, Sprinzl M, Stadler PF, Pütz J. 2009. "tRNAdb 
2009: compilation of tRNA sequences and tRNA genes." Nucleic Acids Research, 
Volume 37 (suppl_1): D159–162. doi:10.1093/nar/gkn772.
}
\author{
Felix G M Ernst [aut]
}
\keyword{datasets}
